massaging templating and routes for data visualizations
