+++ controllers/arShapesController.js
@@ -1,4 +1,3 @@
-const arShapes = require('../models/arShapes');
 const moment = require('moment');
 
 exports.find_one = function(req, res , next) {
