+++ controllers/griddedProductController.js
@@ -1,4 +1,3 @@
-const Grid = require('../models/grid');
 const GridParameter = Grid.ksTempParams;
 const moment = require('moment');
 const helper = require('../public/javascripts/controllers/griddedHelperFunctions')
