+++ package.json
@@ -6,13 +6,8 @@
   },
   "private": true,
   "scripts": {
-    "test": "export NODE_ENV=test && mocha",
     "travis_test": "export NODE_ENV=test && export REMOTE_DB=true && mocha --exit",
-    "start": "export NODE_ENV=production && node ./bin/www",
-    "dockerstart": "export NODE_ENV=docker && node ./bin/www",
-    "devstart": "export NODE_ENV=development && nodemon ./bin/www",
-    "yamldevstart": "export NODE_ENV=development && nodemon ./bin/www --watch ./public/api-docs/swagger.yaml",
-    "jgstart": "export NODE_ENV=jg && nodemon ./bin/www"
+    "start": "export NODE_ENV=production && node ./bin/www"
   },
   "dependencies": {
     "acorn-globals": "^4.3.0",
@@ -26,7 +21,6 @@
     "compression": "^1.7.4",
     "config": "^1.31.0",
     "cookie-parser": "^1.4.4",
-    "cors": "^2.8.5",
     "debug": "3.1.0",
     "express": "^4.16.4",
     "express-validator": "^4.3.0",
@@ -43,7 +37,6 @@
     "leaflet.coordinates": "^0.1.5",
     "leaflet.heat": "^0.2.0",
     "moment": "^2.24.0",
-    "mongoose": ">=5.7.5",
     "morgan": ">=1.9.1",
     "nouislider": "^10.1.0",
     "opencollective": "^1.0.3",
@@ -55,7 +48,6 @@
     "readdirp": "^2.2.1",
     "requirejs": "^2.3.6",
     "serve-favicon": "^2.5.0",
-    "swagger-ui-express": "^4.1.4",
     "tableexport": "^4.0.11",
     "undefsafe": "^2.0.2"
   },
