+++ views/platform_page.pug
@@ -86,7 +86,6 @@ html(lang='en')
 
         $.when(
         $.getScript( "/javascripts/chartFunctions.js" ),
-        $.getScript( "/javascripts/colorBarChartFunctions.js" ),
         $.Deferred(function( deferred ){
             $( deferred.resolve )
         })
