+++ views/platform_page_2.pug
@@ -83,7 +83,6 @@ html(lang='en')
     script(type='text/javascript').
         $.when(
         $.getScript( "/javascripts/chartFunctions.js" ),
-        $.getScript( "/javascripts/scatterChartFunctions.js" ),
         $.Deferred(function( deferred ){
             $( deferred.resolve )
         })
