+++ views/selected_profile_page.pug
@@ -84,7 +84,6 @@ html(lang='en')
     script(type='text/javascript').
         $.when(
         $.getScript( "/javascripts/chartFunctions.js" ),
-        $.getScript( "/javascripts/scatterChartFunctions.js" ),
         $.Deferred(function( deferred ){
             $( deferred.resolve )
         })
